Organize FrameTransformer benchmarks and tests
